+++ app/Models/ProductEnterDetail.php
@@ -10,5 +10,23 @@ class ProductEnterDetail extends Model
 {
     use HasFactory, SoftDeletes;
     protected $table = 'product_enter_detail';
+    protected $fillable = [
+        'product_id',
+        'product_enter-id',
+        'quantity'
+    ];
+
+    /**
+     * Relation
+     */
+
+    public function product()
+    {
+        return $this->belongsTo(Product::class, 'product_id');
+    }
+    public function enter()
+    {
+        return $this->belongsTo(ProductEnter::class, 'product_enter_id');
+    }
 
 }