+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # CHANGE LOG
 
+# 2021-11-06
+
+- Add Product Enter Detail (detail barang masuk)
+- update stock ketika berhasil menambahkan data barang masuk
 # 2021-11-05
+
 - Add Product Enter (proses barang masuk)
+
 # 2021-11-04
 
 - Add Laravolt/Indonesia for Indonesian region data